+++ api/resources/wordpress/client/requests/index.ts
@@ -1,4 +1,3 @@
-export type { ChangePackageDto } from "./ChangePackageDto.js";
 export type { ListPackagesWordpressRequest } from "./ListPackagesWordpressRequest.js";
 export type { ListUpgradePackagesWordpressRequest } from "./ListUpgradePackagesWordpressRequest.js";
 export type { UpdatePasswordDto } from "./UpdatePasswordDto.js";